chore: split path_conversion and generic_path_conversion, remove unused
