+++ ferment/src/item_conversion.rs
@@ -12,7 +12,8 @@ use crate::composer::{ConversionsComposer, ItemComposer};
 use crate::context::Context;
 use crate::formatter::{format_types_dict, format_used_traits};
 use crate::visitor::Visitor;
-use crate::path_conversion::{GenericPathConversion, PathConversion};
+use crate::generic_path_conversion::GenericPathConversion;
+use crate::path_conversion::PathConversion;
 use crate::presentation::{ConversionInterfacePresentation, DocPresentation, DropInterfacePresentation, Expansion, FFIObjectPresentation, FromConversionPresentation, ToConversionPresentation, TraitVTablePresentation};
 use crate::scope::{EMPTY, Scope};
 use crate::import_conversion::{ImportConversion, ImportType};
