+++ ferment/src/visitor.rs
@@ -8,7 +8,8 @@ use crate::Config;
 use crate::context::Context;
 use crate::formatter::{format_types_dict_full, format_used_traits};
 use crate::item_conversion::{ItemContext, ItemConversion};
-use crate::path_conversion::{GenericPathConversion, PathConversion};
+use crate::generic_path_conversion::GenericPathConversion;
+use crate::path_conversion::PathConversion;
 use crate::scope_conversion::ScopeTreeExportItem;
 use crate::scope::Scope;
 use crate::type_conversion::TypeConversion;
